Fix compiler warnings

+++ molecube/parser/parseTxtSeq.cpp
@@ -449,7 +449,8 @@ parseSeqTxt(Pulser::Controller &ctrl, unsigned reps,
         parse_time = toc();
         // Only cache the sequence if it takes longer than 50ms and
         // 20% of the sequence length to parse.
-        if (is_b64 && parse_time > 50e6 && parse_time > builder_p->currT * 2)
+        if (is_b64 && parse_time > (uint64_t)50e6 &&
+            parse_time > builder_p->currT * 2)
             builder_p = seq_cache.set(seqTxt, std::move(_builder));
         if (!builder_p) {
             builder_p = &_builder;
